refactor(react): extract ReAct primitives to src/react.ts to break circular imports
