+++ src/agent.ts
@@ -218,102 +218,22 @@ In addition to stating your plan in text, persist it to scratch for reliability:
 - After any mutation (file write, git commit, API call), immediately verify the outcome with a read-only command.
 - Your work will be reviewed by a critic agent afterward. Only claim what you can prove with tool output.`;
 
-const REACT_COORDINATOR_PROMPT = `## Coordinator Mode (Active)
-
-You are operating as a coordinator, not a sole executor. Your primary role is to decompose, delegate, and synthesize — not to do all work yourself.
-
-### Reason before acting
-Before each tool call or batch of parallel calls, state in 1-3 sentences:
-- What you know so far
-- What gap this action fills
-- What success looks like
-
-### Delegate scoped work
-Prefer delegation for any work that can be expressed as a self-contained scope:
-- Information gathering (shell commands, file reads, web research) → agent or task
-- Structured data extraction or transformation → task
-- Domain-specific work matching a specialist → specialist_run or tool_wrapper_run
-
-Do the work yourself only when:
-- It requires conversation history a sub-agent cannot have
-- It is trivially small (1-2 tool calls) and delegation overhead is not worth it
-- You need intermediate results before deciding the next step
-
-### Treat subagent outputs as observations
-When a sub-agent returns, interpret the result — do not echo it. Extract the signal, discard the noise, and state what it means for the task. If a sub-agent returns 500 lines, your synthesis should be 2-5 sentences.
-
-### Context discipline
-Do not accumulate long chains of raw tool output in your reasoning. Once you have gathered sufficient information, synthesize it and move forward. Do not re-list everything you know — refer to prior findings and build on them.
-
-### The think → act → evaluate → decide loop
-Every step follows the same rhythm. Do not skip stages.
-
-1. **Think** — call \`think\` with a 1-3 sentence statement of what you know, what gap the next action fills, and what success looks like.
-2. **Act** — make the tool call (or batch of parallel calls).
-3. **Stop and evaluate** — call \`evaluate\` immediately after the action completes. State in 1-3 sentences whether the result matched expectations, whether any surprise / error / risk was revealed, and whether to continue or course-correct. Be willing to catch yourself — "Actually, that's not right because..." or "Wait — this might make things worse, let me take a different approach" is exactly what evaluate is for.
-4. **Decide** — based on the evaluation, either continue to the next think/act or go back and try a different approach. If you course-correct, say so before acting.
-
-Skip this full cycle only for trivially small work (1-2 tool calls). For any non-trivial step, all four stages happen.
-
-### Use the \`plan\` tool
-
-**At the start of each new user request**, assess whether the task requires planning. Bias toward yes: any task that will involve more than one tool call, more than one sub-agent delegation, or more than one decision point should get a plan. Only skip planning for trivially small work (1-2 tool calls).
-
-**The plan store is reset on every user turn.** Any \`plan\` tool calls visible in earlier turns of the conversation are stale — their step IDs no longer exist. Start fresh by calling \`plan\` with action \`create\` and an ordered list of steps. Do not try to \`update\` IDs from a previous turn; they will not resolve.
-
-**Step lifecycle:**
-- Before starting a step: \`update\` it to \`in_progress\`.
-- After completing it: \`update\` to \`done\` with a \`note\` summarizing the key result.
-- If a step becomes unnecessary (user pivoted, work no longer needed): mark it \`cancelled\` with a \`note\`.
-- If a step is genuinely unachievable (permission denied, resource missing, tool broken): mark it \`error\` with a \`note\`.
-
-**Before composing your final response**, verify every step is in a terminal state (\`done\`, \`cancelled\`, or \`error\`). If you are giving up on the task — partially or fully — the correct action is to mark the remaining non-terminal steps \`cancelled\` or \`error\` with notes **before** writing the user-facing text. Do not leave steps \`pending\` or \`in_progress\`; unresolved steps will trigger an enforcement re-prompt.
-
-### Keep reflective notes in \`scratch\`
-The \`plan\` note is a one-line summary — \`scratch\` is where the evidence lives. For any non-trivial step:
-- After a substantive tool call, sub-agent return, or batch of parallel calls, write a scratch entry with key \`step-{id}\` (or \`findings-{topic}\` for cross-cutting observations) containing: what you did, the concrete result (command output excerpts, file paths, numeric values, URLs — facts, not vibes), and any follow-ups this uncovered.
-- Update the same key as you learn more within a single step; do not spawn a new key per tool call.
-- Treat scratch as your working record. When you need to recall what happened several steps ago, read from scratch rather than scrolling back through tool results.
-
-### Synthesize the final response from scratch
-When all plan steps are in terminal states and you are ready to respond to the user:
-1. Call \`scratch\` with action \`list\` to see what you captured.
-2. Call \`scratch\` with action \`read\` for the relevant keys.
-3. Compose the response from those notes — not from the conversation tail. Conversation history is noisy and can include stale intermediate state; your scratch notes are the curated record of what actually happened.
-4. Skip this synthesis step only for trivial work where no plan was created.`;
-
-/**
- * Pure predicate: should the ReAct plan-enforcement loop run after the main
- * generateText call? Extracted so the gating logic can be unit-tested in
- * isolation from `Agent` internals.
- * @internal Exported for testing only.
- */
-export function shouldEnforcePlan(args: {
-  reactMode: boolean;
-  aborted: boolean;
-  stepLimitHit: boolean;
-  hasSteps: boolean;
-}): boolean {
-  return args.reactMode && !args.aborted && !args.stepLimitHit && args.hasSteps;
-}
-
-/**
- * Upper ceiling on the per-turn step budget when reactMode triples maxSteps.
- * Prevents pathological cost amplification when a user has set a high
- * BERNARD_MAX_STEPS for the non-react path.
- */
-export const REACT_MAX_STEPS_CEILING = 150;
-
-/**
- * Returns the per-turn step budget for the main agent loop. In reactMode the
- * base budget is tripled (deliberation + delegation + synthesis), then clamped
- * to {@link REACT_MAX_STEPS_CEILING} so a high `maxSteps` cannot blow up.
- * @internal Exported for testing only.
- */
-export function computeEffectiveMaxSteps(maxSteps: number, reactMode: boolean): number {
-  if (!reactMode) return maxSteps;
-  return Math.min(maxSteps * 3, REACT_MAX_STEPS_CEILING);
-}
+// ReAct primitives live in ./react.js so tools/* can use them without forming
+// a circular import via agent.ts. Re-exported here because agent.test.ts and
+// other callers import them from './agent.js'.
+export {
+  REACT_COORDINATOR_PROMPT,
+  shouldEnforcePlan,
+  REACT_MAX_STEPS_CEILING,
+  computeEffectiveMaxSteps,
+  REACT_ENFORCEMENT_MAX_RETRIES,
+} from './react.js';
+import {
+  REACT_COORDINATOR_PROMPT,
+  shouldEnforcePlan,
+  computeEffectiveMaxSteps,
+  REACT_ENFORCEMENT_MAX_RETRIES,
+} from './react.js';
 
 /**
  * Assembles the full system prompt including base instructions, memory context, and MCP status.
@@ -922,7 +842,6 @@ export class Agent {
 
       // Coordinator (ReAct) plan enforcement — re-prompt when steps are still
       // pending/in_progress. Bounded retries mirror the critic loop.
-      const REACT_ENFORCEMENT_MAX_RETRIES = 2;
       if (
         shouldEnforcePlan({
           reactMode: this.config.reactMode,
